+++ modulo3/astromatch/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AllScreens from './components/AllScreens/AllScreens';
+import AllScreens from './components/Home/Home';
 import GlobalStyle from './Styles'
 
 
@@ -7,7 +7,6 @@ const App = () => {
 
   return (
     <div>
-      
       <GlobalStyle/>
       <AllScreens />
     </div>
