--- modulo3/astromatch/src/components/Bottons/BottonsStyles.js
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-
-export const BottonsStyled = styled.img`
-    margin-left: -130px; 
-	height: 6vh;
-    position: absolute;
-	transform: translateX(500%);
-    margin-top: 20px; 
-
-`
