--- modulo3/astromatch/src/components/Footer/footerStyles.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const FooterStyled = styled.div`
-    /* display: flex;
-    flex-direction: column; */
-    /* height: 89px;
-    position: fixed;
-    border-radius: 40px;
-    box-shadow: rgb(0 0 0 / 30%) 0px 0px 50px;
-    border-top-right-radius: 0px;
-    border-top-left-radius: 0px; */
-`;
