+++ modulo3/astromatch/src/components/Home/Home.js
@@ -1,0 +1,16 @@
+import React from 'react';
+// import Load from '../Load/Load';
+import * as C from './HomeStyles'
+
+const AllScreens = () => {
+
+  return (
+    <div>
+      <C.ContainerHome>
+      {/* <Load /> */}
+      </C.ContainerHome>
+    </div>
+  );
+};
+
+export default AllScreens;
