+++ modulo3/astromatch/src/components/Home/HomeStyles.js
@@ -1,16 +1,14 @@
 import styled from "styled-components";
 
-export const ContainerAllScreens = styled.div`
+export const ContainerHome = styled.div`
     display: flex;
     flex-direction: column;
-    /* width: 27vw;
-    height: 85vh; */
-    width: 350px;
+    width: 300px;
     height: 85vh;
     position: fixed;
-    margin: 6vh  37%;
+    margin: 6vh  35%;
     border-radius: 40px;
-    background-color: #5F9EA0;
-     /* background-color: #CD5C5C; */
+    /* background-color: #5F9EA0; */
+     background-color: #CD5C5C;
     box-shadow: rgb(0 0 0 / 50%) 0px 0px 8px;
 `;
