+++ modulo3/astromatch/src/components/Load/Load.js
@@ -1,0 +1,3 @@
+import React, {useState, useEffect} from "react";
+import * as C from './LoadStyles'
+import axios from "axios";
