+++ modulo3/astromatch/src/components/Load/LoadStyles.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+
