--- modulo3/astromatch/src/components/MainScreen/MainsScreenStyles.js
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-export const MainScreenStyled = styled.div`
-    display: flex;
-    /* margin-top: 0;
-    height: 395px;
-    width: 350px;
-    position: absolute;
-    background-color: white;  */
-`;
