--- modulo3/astromatch/src/components/TextMainScreen/TextMainStyles.js
@@ -1,31 +1,0 @@
-import styled from "styled-components";
-
-export const TextMainScreenStyled = styled.div`
-
-        h3 {
-                margin-top: 308px; 
-                color: white; 
-                position:relative ;
-                margin-left: 50px; 
-                
-
-        }
-        p {
-                display: flex;
-                align-items: center; 
-                flex-direction: column;
-                color: white; 
-                margin-left: 20px; 
-                margin-right: 20px; 
-                position:relative ;
-                font-weight: 500px;
-        }
-
-        img { 
-        display: flex;
-        margin-top: 0px;
-        height: 400px;
-        width: 350px;
-        position:fixed;
-        }        
-        `
