+++ modulo3/astromatch/src/components/header/header.js
@@ -1,12 +1,34 @@
-import React from "react";
-import * as C from './headerStyles'
+import React, {useState, useEffect} from "react";
+import Buttons from "../Buttons/Buttons";
+import * as C from './HeaderStyles'
+import axios from "axios";
 
-const Header = () => {
+
+const TextMainScreen = () => {
+  const [profile, setProfile] = useState([])
+
+  const getProfiles = () => {
+    axios.get('https://us-central1-missao-newton.cloudfunctions.net/astroMatch/gus/person')
+    .then((response)=> {
+      console.log(response.data);
+      setProfile(response.data.profile)
+    }).catch((error)=>{
+      console.log(error.message)
+    })
+  }
+
+    useEffect (()=>{
+      getProfiles()
+    }, [])
 
     return (
-            <C.HeaderStyled>
-            <h2>Astromatch</h2>    
-            </C.HeaderStyled>
-    )
-}
-export default Header;
+    <C.TextMainScreenStyled>
+    <img src={profile.photo} alt={profile.name} />
+    <h3> 🟢{profile.name}, {profile.age}</h3> 
+    <p>{profile.bio}</p>
+    <Buttons id={profile.id}/>
+    </C.TextMainScreenStyled> 
+    );
+  };
+  
+  export default TextMainScreen;
